+++ src/django_components/commands/components.py
@@ -6,16 +6,16 @@ from django_components.util.command import ComponentCommand
 
 
 # TODO_V3 - This command should be called standalone as "components":
-#           `python -m components start <name>`
-#           `components start <name>`
+#           `python -m components create <name>`
+#           `components create <name>`
 class ComponentsRootCommand(ComponentCommand):
     """
     The entrypoint for the "components" commands.
 
     ```bash
     python manage.py components list
-    python manage.py components start <name>
-    python manage.py components upgrade <name>
+    python manage.py components create <name>
+    python manage.py components upgrade
     python manage.py components ext list
     python manage.py components ext run <extension> <command>
     ```